requirements.txt
requests>=2.25.1
ipwhois>=1.2.0
requests==2.31.0
dnspython==2.0.0
tqdm==4.66.1
tkhtmlview==0.3.1
ttkbootstrap==1.10.1
